+++ lqcv/data/converter/base.py
@@ -1,0 +1,93 @@
+from abc import ABCMeta, abstractmethod
+from lqcv.utils.log import LOGGER
+from lqcv.utils.plot import plot_one_box, colors
+from collections import defaultdict
+from tqdm import tqdm
+from tabulate import tabulate
+import os.path as osp
+import cv2
+
+
+
+class BaseConverter(metaclass=ABCMeta):
+    def __init__(self, label_dir, class_names=None, img_dir=None) -> None:
+        super().__init__()
+        self.labels = list()
+        self.catCount = defaultdict(int)
+        self.catImgCnt = dict()
+        self.imgs_wh = dict()
+        self.img_dir = img_dir
+        self.class_names = class_names
+
+        self.read_labels(label_dir)
+
+    # @abstractmethod
+    # def toCOCO(self):
+    #     pass
+    #
+    # @abstractmethod
+    # def toXML(self):
+    #     pass
+    #
+    # @abstractmethod
+    # def toYOLO(self):
+    #     pass
+
+    @abstractmethod
+    def read_labels(self, label_dir):
+        pass
+
+    def visualize(self, save_dir=None):
+        if self.img_dir is None:
+            LOGGER.warning("`self.img_dir` is None.")
+            return
+
+        pbar = tqdm(self.labels, total=len(self.labels))
+        if save_dir is None:
+            cv2.namedWindow("p", cv2.WINDOW_NORMAL)
+        for label in pbar:
+            try:
+                filename = label["img_name"]
+                image = cv2.imread(osp.join(self.img_dir, filename))
+                if image is None:
+                    continue
+                cls, bbox = label["cls"], label["bbox"]
+                bbox.convert("xyxy")
+                for i, c in enumerate(cls):
+                    plot_one_box(
+                        bbox.data[i],
+                        image,
+                        color=colors(int(c)),
+                        line_thickness=2,
+                        label=self.class_names[int(c)],
+                    )
+            except Exception as e:
+                LOGGER.warning(e)
+                continue
+            if save_dir is not None:
+                cv2.imwrite(osp.join(save_dir, filename), image)
+            else:
+                cv2.imshow("p", image)
+                if cv2.waitKey(0) == ord("q"):
+                    break
+
+    def __repr__(self):
+        total_img = 0
+        total_obj = 0
+        cat_table = []
+        for i, c in enumerate(self.class_names):
+            if c in self.catImgCnt and c in self.catCount:
+                total_img += self.catImgCnt[c]
+                total_obj += self.catCount[c]
+                cat_table.append((str(i), str(c), self.catImgCnt[c], self.catCount[c]))
+            else:
+                cat_table.append((str(i), str(c), 0, 0))
+
+        cat_table += [(" ", "total", total_img, total_obj)]
+        return "\n" + tabulate(
+            cat_table,
+            headers=["Id", "Category", "ImageCnt", "ClassCnt"],
+            tablefmt="fancy_grid",
+            missingval="None",
+        )
+
